:recycle: Complete refactoring core
