+++ src/containers/app/types.ts
@@ -1,5 +1,6 @@
+import { SupportedIcon } from "@core/hooks/use-icon";
 import { InternalSettings, UserSettings } from "@core/settings/types";
-import { OrdoEvent, SupportedIcon } from "@core/types";
+import { OrdoEvent } from "@core/types";
 import { OrdoEvents } from "@init/types";
 
 export type Command<TCustomEvents extends Record<string, any> | null = null> = {
