+++ src/core/types.ts
@@ -1,5 +1,3 @@
-import * as HiIcons from "react-icons/hi";
-
 import { EventHandler } from "@core/transmission";
 import { OrdoEvents } from "@init/types";
 
@@ -16,5 +14,3 @@ export type OrdoEvent<TScope extends string = string, TEvent extends string = st
   `@${TScope}/${TEvent}`,
   TPayload
 >;
-
-export type SupportedIcon = keyof typeof HiIcons;
